+++ src/css/home.css
@@ -7,7 +7,7 @@
   ;
 }
 .logoHome {
-  margin-top: 40%;;
+  margin-top: 40%;
   width: 50%;
 }
 .containerHome {
